# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayTrace Backend",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/evaluation-runs": {
      "post": {
        "tags": [
          "evaluations",
          "evaluations"
        ],
        "summary": "Create Evaluation",
        "operationId": "create_evaluation_api_v1_evaluation_runs_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunTriggerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "evaluations",
          "evaluations"
        ],
        "summary": "List Evaluations",
        "operationId": "list_evaluations_api_v1_evaluation_runs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluation-runs/{evaluation_run_id}": {
      "get": {
        "tags": [
          "evaluations",
          "evaluations"
        ],
        "summary": "Get Evaluation",
        "operationId": "get_evaluation_api_v1_evaluation_runs__evaluation_run_id__get",
        "parameters": [
          {
            "name": "evaluation_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluation Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/evaluation-runs/{evaluation_run_id}/report": {
      "get": {
        "tags": [
          "evaluations",
          "evaluations"
        ],
        "summary": "Download Evaluation Report",
        "operationId": "download_evaluation_report_api_v1_evaluation_runs__evaluation_run_id__report_get",
        "parameters": [
          {
            "name": "evaluation_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Evaluation Run Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "markdown"
              ],
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AggregateMetrics": {
        "properties": {
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count"
          },
          "succeeded_count": {
            "type": "integer",
            "title": "Succeeded Count"
          },
          "run_success_rate": {
            "type": "number",
            "title": "Run Success Rate"
          },
          "stage_localization_exact_rate": {
            "type": "number",
            "title": "Stage Localization Exact Rate"
          },
          "stage_localization_overlap_mean": {
            "type": "number",
            "title": "Stage Localization Overlap Mean"
          },
          "root_cause_precision_mean": {
            "type": "number",
            "title": "Root Cause Precision Mean"
          },
          "root_cause_recall_mean": {
            "type": "number",
            "title": "Root Cause Recall Mean"
          },
          "root_cause_f1_mean": {
            "type": "number",
            "title": "Root Cause F1 Mean"
          },
          "evidence_validity_rate_mean": {
            "type": "number",
            "title": "Evidence Validity Rate Mean"
          },
          "unsupported_claim_rate_mean": {
            "type": "number",
            "title": "Unsupported Claim Rate Mean"
          },
          "loss_attribution_mae": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loss Attribution Mae"
          },
          "tool_call_count_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Count Mean"
          },
          "latency_ms_mean": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms Mean"
          },
          "badcase_count": {
            "type": "integer",
            "title": "Badcase Count"
          },
          "total_input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Input Tokens"
          },
          "total_output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Output Tokens"
          },
          "estimated_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost"
          }
        },
        "type": "object",
        "required": [
          "scenario_count",
          "succeeded_count",
          "run_success_rate",
          "stage_localization_exact_rate",
          "stage_localization_overlap_mean",
          "root_cause_precision_mean",
          "root_cause_recall_mean",
          "root_cause_f1_mean",
          "evidence_validity_rate_mean",
          "unsupported_claim_rate_mean",
          "badcase_count"
        ],
        "title": "AggregateMetrics",
        "description": "Summary metrics across all scenarios in one EvaluationRun (\u00a7 20.2)."
      },
...
      "EvaluationRunCreate": {
        "properties": {
          "model_mode": {
            "type": "string",
            "enum": [
              "B0"
            ],
            "const": "B0",
            "title": "Model Mode",
            "default": "B0"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version",
            "default": "rule-based.v1"
          },
          "scenario_kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Scenario Kinds"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "num_intents": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Num Intents",
            "default": 500
          }
        },
        "type": "object",
        "title": "EvaluationRunCreate"
      },
      "EvaluationRunListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvaluationRunResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "EvaluationRunListResponse"
      },
      "EvaluationRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "model_mode": {
            "type": "string",
            "title": "Model Mode"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "ontology_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ontology Version"
          },
          "scenario_kinds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scenario Kinds"
          },
          "scenario_count": {
            "type": "integer",
            "title": "Scenario Count"
          },
          "seed": {
            "type": "integer",
            "title": "Seed"
          },
          "num_intents": {
            "type": "integer",
            "title": "Num Intents"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AggregateMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "scenario_results": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ScenarioResult"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Results"
          },
          "badcases": {
            "anyOf": [
              {
                "items": {
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Badcases"
          },
          "report_json_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Json Key"
          },
          "report_markdown_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Markdown Key"
          },
          "error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Type"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "celery_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Celery Task Id"
          },
          "total_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration Ms"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "idempotency_key",
          "status",
          "model_mode",
          "model_name",
          "prompt_version",
          "ontology_version",
          "scenario_kinds",
          "scenario_count",
          "seed",
          "num_intents",
          "metrics",
          "scenario_results",
          "badcases",
          "report_json_key",
          "report_markdown_key",
          "error_type",
          "error_message",
          "celery_task_id",
          "total_duration_ms",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at"
        ],
        "title": "EvaluationRunResponse"
      },
      "EvaluationRunTriggerResponse": {
        "properties": {
          "evaluation_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evaluation Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "evaluation_run_id",
          "status",
          "message"
        ],
        "title": "EvaluationRunTriggerResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScenarioResult": {
        "properties": {
          "scenario_kind": {
            "type": "string",
            "title": "Scenario Kind"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "diagnosis_status": {
            "type": "string",
            "title": "Diagnosis Status"
          },
          "predicted_anomalous_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predicted Anomalous Stages"
          },
          "expected_anomalous_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Anomalous Stages"
          },
          "predicted_root_causes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predicted Root Causes"
          },
          "expected_root_causes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Root Causes"
          },
          "predicted_missing_data": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Predicted Missing Data"
          },
          "expected_data_gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Expected Data Gaps"
          },
          "stage_exact": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Exact"
          },
          "stage_overlap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Overlap"
          },
          "root_cause_precision": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Precision"
          },
          "root_cause_recall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause Recall"
          },
          "root_cause_f1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause F1"
          },
          "evidence_validity_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Validity Rate"
          },
          "unsupported_claim_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unsupported Claim Rate"
          },
          "loss_attribution_abs_error": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loss Attribution Abs Error"
          },
          "tool_call_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Count"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "report_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Summary"
          },
          "unexplained_lost_intents": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unexplained Lost Intents"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "evidence": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "tool_trace": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Tool Trace"
          },
          "badcases": {
            "items": {
              "type": "string",
              "enum": [
                "STAGE_MISS",
                "STAGE_FALSE_POSITIVE",
                "ROOT_CAUSE_MISS",
                "ROOT_CAUSE_FALSE_POSITIVE",
                "UNSUPPORTED_CLAIM",
                "LOSS_ATTRIBUTION_ERROR",
                "RUN_FAILURE",
                "DATA_GAP_MISS"
              ]
            },
            "type": "array",
            "title": "Badcases"
          }
        },
        "type": "object",
        "required": [
          "scenario_kind",
          "scenario_id",
          "diagnosis_status"
        ],
        "title": "ScenarioResult",
        "description": "Per-scenario evaluation outcome (prediction vs Ground Truth)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}